Matches design
Label.js:
css custom properties are now prefixed with label-- to
prevent collisions with other components. This is
because the CSS now sits on :root instead of ul-label.
This allows for containing elements to set avengers
theme or size setting for a whole segment of the UI
without having to repeat that in all markup.

screenshots/*/*:
Added images of components from figma for easy
comparison of design against code.

styles/styles.css:
updated layout of demo

themes/*.css
Added themes which just define a set of custom
variables. These are set on :root, which allows them to
be applied anywhere in the document hierarchy
which makes it possible to set an entire region as
being avengers style or a certain size.