more robust styling, by making the label built from divs
The label is now made from 3 divs.
1. the icon
2. the text label
3. the close button

The label uses flex to layout contents.
The icon and close button are limited to the height
of the label, but the text is allowed to wrap.

The outer layout aligns then flex-start, which keeps
the icon and close button on the first line if it wraps.