fix(Label): :bug: Populate method now rebuilds component completely to avoid side effects.
If a component was removed, and then readded to the document
it would end up with multiple copies of the icon, text and close
button. This is because adding it back causes populate to be called
again, with no object state but with the shadow dom populated as
before. This caused it to add another icon, text, etc.

This removes and rebuilds the visible content (but not
the style tag, since that's done in the constructor) which
addresses the problem.

Also in html example, updating when the label has been dismissed
is not caught and doesn't error.